WIP: impl useSearchOperation
